BUG FIX: Fixed Killer DTO, Controller and Mapper

+++ src/main/java/com/example/human_vs_zombies/services/squadCheckIn/SquadCheckInServiceIml.java
@@ -29,8 +29,9 @@ public class SquadCheckInServiceIml implements SquadCheckInService {
     @Override
     public void deleteById(Integer integer) {
         //check if exists
-        this.findById(integer);
-        //delete
-        squadCheckInRepository.deleteById(integer);
+        if(squadCheckInRepository.existsById(integer)){
+            squadCheckInRepository.deleteById(integer);
+        }
+
     }
 }
